rename shipping bags folder
