+++ packages/js/product-editor/src/images/shopping-bags/index.tsx
