+++ packages/js/product-editor/src/images/shopping-bags/stories/index.tsx
